refactor : make vo in auth
