+++ src/main/java/spaland/auth/AuthenticationService.java
@@ -8,10 +8,11 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import spaland.auth.vo.RefreshRequest;
+import spaland.auth.vo.SignupRequest;
 import spaland.config.JwtService;
 import spaland.email.service.RedisService;
 import spaland.error.ApiException;
-import spaland.error.ErrorCode;
 import spaland.users.model.Role;
 import spaland.users.model.User;
 import spaland.users.repository.IUserRepository;
