--- src/main/java/spaland/auth/RefreshRequest.java
@@ -1,14 +1,0 @@
-package spaland.auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class RefreshRequest {
-    private String refreshToken;
-}
